+++ components/Masthead/Masthead.styles.ts
@@ -10,4 +10,4 @@ export const lockup = 'relative top-2 sm:top-0 z-[300] sm:-mt-02em shrink-0 sm:t
 export const flexbox = 'gap-x-12 sm:gap-x-20 md:gap-x-30';
 
 // We use the group-has- TW selector to make the Giving link white when the main nav is open
-export const cta = 'relative z-[300] group-has-[nav[data-headlessui-state="open"]]:text-white group-has-[nav[data-headlessui-state="open"]]:decoration-white'
+export const cta = 'relative z-[300] group-has-[nav[data-headlessui-state="open"]]:text-white group-has-[nav[data-headlessui-state="open"]]:decoration-white';
